ppt-translator/pyproject.toml
[project]
dependencies = [
    "python-pptx",
    "deep-translator"
]
name = "ppt-translator"
requires-python = ">= 3.11"
version = "0.1.0"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["win-64"]

[tool.pixi.pypi-dependencies]
ppt_translator = { path = ".", editable = true }

[tool.pixi.tasks]
